Export arrays as colon delimited, and splitenv PATH env vars
Prior to this fix, fish would export arrays as ASCII record separator
delimited, except for a whitelist (PATH, CDPATH, MANPATH). This is
surprising and awkward for other programs to deal with, and there's no way
to get similar behavior for other variables like GOPATH or LD_LIBRARY_PATH.

This commit does the following:

1. Exports all arrays as colon delimited strings, instead of RS.

2. When importing environment variables, if the variable ends in PATH,
split it on colons (i.e. splitenv); otherwise do not split it.

Colons are not escaped; this is deliberate to support uses like

   `set -x PYTHONPATH "/foo:/bar"`

which ought to work (and already do, we don't want  to make a compat break
here).

Fixes (or at least mitigates) #436