Variables to explicitly track their delimiters
This equips fish variables with an explicitly set delimiter. This delimiter
is ':' for PATH variables and ' ' (space) for other variables.

Delimiters are attached to variables, not to variable names. It is possible
for a local variable to have a different delimiter than a variable it shadows.

When creating a new variable, delimiters are inferred from the existing
variable (perhaps in a parent scope). This allows e.g. locally set PYTHONPATH
to infer the delimiter from any global PYTHONPATH.

Delimiters are not currently used.