Bravely remove reclaim_foreground_pgrp param from continue_job
The original idea here was to avoid SIGTTINs by only transferring the
tty ownership if fish was already the owner. This dated from a time
before we had really nailed down how pgroups should be assigned.
Now we more deliberately assign a job's pgroup so we don't need
this conservative check.