Allow timing-out I/O-able syntax highlighting after expanding abbreviation
It may happen that the user types an abbreviation and then hits return.
Prior to this commit, we would perform a form of syntax highlighting
that does not require I/O, so as to not block the user. However this
could cause invalid commands to be colored as valid.

Introduce a new function which performs background I/O and then waits
for a little bit (250 milliseconds) for it to complete. If it does not
complete in time, fall back to the no-I/O variant.

This fixes #7418; a followup commit will address the more general #5912.