Adopt imstring in lots of places
This switches many use cases from `wcstring` to `imstring`, which saves
memory.

There's two types of change:

1. Functions which accepted a `const wcstring &` should now generally
accept an `const imstring &` instead. For example, `escape_string` does
this. Because the APIs are broadly compatible, and a `const imstring &`
may be constructed from a `wcstring`, this is usually as easy as just
changing the string type.

2. Classes which store strings which do not change. For example, variable
stack now uses `imstring` for variable names. This allows them to be
constructed from string literals without copying.

The change is necessarily large because it cascades over - switching to
imstring causes many other places to switch to `imstring` as well.