Introduce imstring
imstring is a new string class intended to address the problems of
wcstring. imstring is immutable, and multiple imstrings may share the same
backing storage.

In almost every case, a `const wcstring &` parameter should be replaced by
`const imstring &`, because when the `imstring` is constructed from a
string literal it will not copy the data. Likewise `const wchar_t *` params
can start being replaced by `const imstring &`.

An imstring may be in one of three states:

1. Referencing unowned data, which it will copy when the `imstring` itself
is copied or moved.
2. Referencing owned data in a `shared_ptr`. Here the storage is
reference-counted.
3. Referencing immortal data like a string literal, which will never be
copied.

imstring subclasses `string_view` so functions like `find_first_of`, etc.
are available. This makes it a drop-in replacement for `wcstring` in many
places. The next commit starts adopting it.