Add "path" builtin
This adds a "path" builtin that can handle paths.

Implemented so far:

- "path expand GLOBS", which will run glob(3) on the paths
- "path filter PATHS", filters paths according to existence and optionally type and permissions
- "path base" and "path dir", run basename and dirname, respectively
- "path extension PATHS", prints the extension, if any
- "path strip-extension", prints the path without the extension
- "path normalize PATHS", normalizes paths - removing "/./" components
- and such.
- "path real", does realpath - i.e. normalizing *and* link resolution.

Some of these - base, dir, {strip-,}extension and normalize operate on the paths only as strings, so they handle nonexistent paths. match, filter and real ignore any nonexistent paths.

All output is split explicitly, so paths with newlines in them are
handled correctly. Alternatively, all subcommands have a "--null-input"/"-z" and "--null-output"/"-Z" option to handle null-terminated input and create null-terminated output. So

    find . -print0 | path base -z

prints the basename of all files in the current directory,
recursively.

With "-Z" it also prints it null-separated.

(if stdout is going to a command substitution, we probably want to
skip this)

All subcommands also have a "-q"/"--quiet" flag that tells them to skip output. They return true "when something happened". For match/filter that's when a file passed, for "base"/"dir"/"extension"/"strip-extension" that's when something about the path *changed*.

Filtering
---------

Filtering works both for `match` (after globbing) and `filter`. It supports all the file*types* `test` has - "dir", "file", "link", "block"..., as well as the permissions - "read", "write", "exec" and things like "suid".

It is missing the tty check and the check for the file being non-empty. The former is best done via `isatty`, the latter I don't think I've ever seen used.

There currently is no way to only get "real" files, i.e. ignore links pointing to files.

Globs
-----

As said above, `path match` runs the arguments as globs, and returns all the matching paths. Directories will have a trailing "/" automatically to indicate it.

Because it uses `glob(3)`, it gets `[a-z]` ranges/classes (with `[!a]` inverted classes and `[[:lower:]]` defined classes) and `?` single-character globs for free.

A few issues with that approach:

1. has extensions on some platforms, so `[^0-9]` works with glibc but will fail on other systems.
2. doesn't do `**`, so we error out on purpose if we detect it, so we can add it later.

Examples
--------

> path match '/usr/bin/?[ai]sh' -Z | path filter -Zz --perm=exec --type=file | path base -z
bash
dash
fish

> path real /bin///sh
/usr/bin/bash

> path extension foo.mp4
mp4

> path extension ~/.config
  (nothing, because ".config" isn't an extension.)