Fix NetBSD executable path to not use procfs
netbsd can sometime give an absolute path but with an odd period and extra slash: for example:

"/path/to/./executable"

to fix this, i would recommend using realpath, or writing your own realpath implementation based on the one netbsd uses in its source code, but without the PATH_MAX limitation. Let me know what your thoughts are on this.