fish_git_prompt: check for xcrun_db size along with existence
Prior to this change, to detect if git was ready on macOS, we detected
the existence of the database file. However xcrun will create the file
immediately, and only populate it later; therefore the file may exist
even though git is not ready. This leads to a slowdown at the git
prompt.

To fix this, check for the size of the file instead. The file appears to
be initially 16 bytes; if larger then that assume git is ready.