Stop requiring Send+Sync from debounce completion callbacks
Today, debounce-style work items are only created from the main thread.
The work to compute the result is done in a background thread but the
completion callback is called on the main thread again.

The completion callbacks used by the reader capture a shared reference to
ReaderData, which includes a Parser.  Neither of those types needs to be
sent across threads.

The debounce machinery moves the completion callback into a function object
that is moved to the background thread and back again.  Because of this
there is a Send+Sync requirement on the completion callback, even though
it's never called on the background thread.

Let's keep the completion callback on the main thread, dropping the need
for Send+Sync.