Make some of the scoped_push stuff nicer
We often use the quite wordy "scoped_push_replacer_ctx" which then becomes the
new "self" which is written as "zelf." But many of these cases are just
twiddling a boolean in the parser's library data.

Make this cleaner by making the library data Rc. This allows the scoped_push to
have its OWN reference to it, so we no longer need the 'zelf' sillyness in these
cases.

Add a type Cleanup which is like ScopedPush except must_use and also can't be
dereferenced. This is to avoid thinking that we need to derefrence through this
return value.

Begin adding a family of functions to handle stuff like suppressing fish_trace.