Make the transient command line a single value, not a stack
This concerns completions which wrap other completions. For example, if 'tig'
wraps 'git' then for the duration of the git completions, we need to make a fake
("transient") command line which contains "git" and not "tig".

Previously we had a stack of such command lines, but we never inspected anything
except the last element. Make this a single value.